@@ -31,6 +31,10 @@ import (
 	"github.com/docker/docker/pkg/stdcopy"
 	"github.com/google/go-cmp/cmp"
 	"github.com/pkg/errors"
+	"gopkg.in/src-d/go-git.v4"
+
+	"github.com/buildpacks/pack/internal/config"
+	"github.com/buildpacks/pack/internal/dist"
 
 	"github.com/buildpacks/pack/internal/archive"
 	"github.com/buildpacks/pack/internal/stringset"
@@ -54,10 +58,12 @@ func AssertEq(t *testing.T, actual, expected interface{}) {
 }
 
 func AssertTrue(t *testing.T, actual interface{}) {
+	t.Helper()
 	AssertEq(t, actual, true)
 }
 
 func AssertFalse(t *testing.T, actual interface{}) {
+	t.Helper()
 	AssertEq(t, actual, false)
 }
 
@@ -102,6 +108,36 @@ func AssertContains(t *testing.T, actual, expected string) {
 	}
 }
 
+func AssertContainsAllInOrder(t *testing.T, actual bytes.Buffer, expected ...string) {
+	t.Helper()
+
+	var tested []byte
+
+	for _, exp := range expected {
+		b, found := readUntilString(&actual, exp)
+		tested = append(tested, b...)
+
+		if !found {
+			t.Fatalf("Expected '%s' to include all of '%s' in order", string(tested), strings.Join(expected, ", "))
+		}
+	}
+}
+
+func readUntilString(b *bytes.Buffer, expected string) (read []byte, found bool) {
+	for {
+		s, err := b.ReadBytes(expected[len(expected)-1])
+		if err != nil {
+			return append(read, s...), false
+		}
+
+		read = append(read, s...)
+		if bytes.HasSuffix(read, []byte(expected)) {
+			return read, true
+		}
+	}
+}
+
+// AssertContainsMatch matches on content by regular expression
 func AssertContainsMatch(t *testing.T, actual, exp string) {
 	t.Helper()
 	if !hasMatches(actual, exp) {
@@ -128,6 +164,14 @@ func AssertSliceContains(t *testing.T, slice []string, expected ...string) {
 	_, missing, _ := stringset.Compare(slice, expected)
 	if len(missing) > 0 {
 		t.Fatalf("Expected %s to contain elements %s", slice, missing)
+	}
+}
+
+func AssertSliceNotContains(t *testing.T, slice []string, expected ...string) {
+	t.Helper()
+	_, missing, _ := stringset.Compare(slice, expected)
+	if len(missing) != len(expected) {
+		t.Fatalf("Expected %s not to contain elements %s", slice, expected)
 	}
 }
 
@@ -272,8 +316,8 @@ func Eventually(t *testing.T, test func() bool, every time.Duration, timeout tim
 func CreateImage(t *testing.T, dockerCli client.CommonAPIClient, repoName, dockerFile string) {
 	t.Helper()
 
-	buildContext, err := archive.CreateSingleFileTarReader("Dockerfile", dockerFile)
-	AssertNil(t, err)
+	buildContext := archive.CreateSingleFileTarReader("Dockerfile", dockerFile)
+	defer buildContext.Close()
 
 	resp, err := dockerCli.ImageBuild(context.Background(), buildContext, dockertypes.ImageBuildOptions{
 		Tags:           []string{repoName},
@@ -283,14 +327,15 @@ func CreateImage(t *testing.T, dockerCli client.CommonAPIClient, repoName, docke
 	})
 	AssertNil(t, err)
 
-	err = checkResponse(resp)
+	defer resp.Body.Close()
+	err = checkResponse(resp.Body)
 	AssertNil(t, errors.Wrapf(err, "building image %s", style.Symbol(repoName)))
 }
 
 func CreateImageFromDir(t *testing.T, dockerCli client.CommonAPIClient, repoName string, dir string) {
 	t.Helper()
 
-	buildContext := archive.ReadDirAsTar(dir, "/", 0, 0, -1, true)
+	buildContext := archive.ReadDirAsTar(dir, "/", 0, 0, -1, true, nil)
 	resp, err := dockerCli.ImageBuild(context.Background(), buildContext, dockertypes.ImageBuildOptions{
 		Tags:           []string{repoName},
 		Remove:         true,
@@ -299,15 +344,15 @@ func CreateImageFromDir(t *testing.T, dockerCli client.CommonAPIClient, repoName
 	})
 	AssertNil(t, err)
 
-	err = checkResponse(resp)
+	defer resp.Body.Close()
+	err = checkResponse(resp.Body)
 	AssertNil(t, errors.Wrapf(err, "building image %s", style.Symbol(repoName)))
 }
 
-func checkResponse(response dockertypes.ImageBuildResponse) error {
-	defer response.Body.Close()
-	body, err := ioutil.ReadAll(response.Body)
+func checkResponse(responseBody io.Reader) error {
+	body, err := ioutil.ReadAll(responseBody)
 	if err != nil {
-		return err
+		return errors.Wrap(err, "reading body")
 	}
 
 	messages := strings.Builder{}
@@ -361,12 +406,16 @@ func DockerRmi(dockerCli client.CommonAPIClient, repoNames ...string) error {
 func PushImage(dockerCli client.CommonAPIClient, ref string, registryConfig *TestRegistryConfig) error {
 	rc, err := dockerCli.ImagePush(context.Background(), ref, dockertypes.ImagePushOptions{RegistryAuth: registryConfig.RegistryAuth()})
 	if err != nil {
-		return err
+		return errors.Wrap(err, "pushing image")
 	}
-	if _, err := io.Copy(ioutil.Discard, rc); err != nil {
-		return err
+
+	defer rc.Close()
+	err = checkResponse(rc)
+	if err != nil {
+		return errors.Wrap(err, "push response")
 	}
-	return rc.Close()
+
+	return nil
 }
 
 func HTTPGetE(url string, headers map[string]string) (string, error) {
@@ -502,8 +551,8 @@ func RecursiveCopy(t *testing.T, src, dst string) {
 }
 
 func RequireDocker(t *testing.T) {
-	_, isSet := os.LookupEnv("NO_DOCKER")
-	SkipIf(t, isSet, "Skipping because docker daemon unavailable")
+	noDocker := os.Getenv("NO_DOCKER")
+	SkipIf(t, strings.ToLower(noDocker) == "true" || noDocker == "1", "Skipping because docker daemon unavailable")
 }
 
 func SkipIf(t *testing.T, expression bool, reason string) {
@@ -577,6 +626,143 @@ func writeTAR(t *testing.T, srcDir, tarDir string, mode int64, w io.Writer) {
 	tw := tar.NewWriter(w)
 	defer tw.Close()
 
-	err := archive.WriteDirToTar(tw, srcDir, tarDir, 0, 0, mode, true)
+	err := archive.WriteDirToTar(tw, srcDir, tarDir, 0, 0, mode, true, nil)
 	AssertNil(t, err)
+}
+
+func RecursiveCopyNow(t *testing.T, src, dst string) {
+	t.Helper()
+	err := os.MkdirAll(dst, 0755)
+	AssertNil(t, err)
+
+	fis, err := ioutil.ReadDir(src)
+	AssertNil(t, err)
+	for _, fi := range fis {
+		if fi.Mode().IsRegular() {
+			srcFile, err := os.Open(filepath.Join(src, fi.Name()))
+			AssertNil(t, err)
+			dstFile, err := os.Create(filepath.Join(dst, fi.Name()))
+			AssertNil(t, err)
+			_, err = io.Copy(dstFile, srcFile)
+			AssertNil(t, err)
+			modifiedTime := time.Now().Local()
+			err = os.Chtimes(filepath.Join(dst, fi.Name()), modifiedTime, modifiedTime)
+			AssertNil(t, err)
+			err = os.Chmod(filepath.Join(dst, fi.Name()), 0664)
+			AssertNil(t, err)
+		}
+		if fi.IsDir() {
+			err = os.Mkdir(filepath.Join(dst, fi.Name()), fi.Mode())
+			AssertNil(t, err)
+			RecursiveCopyNow(t, filepath.Join(src, fi.Name()), filepath.Join(dst, fi.Name()))
+		}
+	}
+	modifiedTime := time.Now().Local()
+	err = os.Chtimes(dst, modifiedTime, modifiedTime)
+	AssertNil(t, err)
+	err = os.Chmod(dst, 0775)
+	AssertNil(t, err)
+}
+
+func AssertTarFileContents(t *testing.T, tarfile, path, expected string) {
+	t.Helper()
+	exist, contents := tarFileContents(t, tarfile, path)
+	if !exist {
+		t.Fatalf("%s does not exist in %s", path, tarfile)
+	}
+	AssertEq(t, contents, expected)
+}
+
+func tarFileContents(t *testing.T, tarfile, path string) (exist bool, contents string) {
+	t.Helper()
+	r, err := os.Open(tarfile)
+	AssertNil(t, err)
+	defer r.Close()
+
+	tr := tar.NewReader(r)
+	for {
+		header, err := tr.Next()
+		if err == io.EOF {
+			break
+		}
+		AssertNil(t, err)
+
+		if header.Name == path {
+			buf, err := ioutil.ReadAll(tr)
+			AssertNil(t, err)
+			return true, string(buf)
+		}
+	}
+	return false, ""
+}
+
+func AssertTarHasFile(t *testing.T, tarFile, path string) {
+	t.Helper()
+
+	exist := tarHasFile(t, tarFile, path)
+	if !exist {
+		t.Fatalf("%s does not exist in %s", path, tarFile)
+	}
+}
+
+func tarHasFile(t *testing.T, tarFile, path string) (exist bool) {
+	t.Helper()
+
+	r, err := os.Open(tarFile)
+	AssertNil(t, err)
+	defer r.Close()
+
+	tr := tar.NewReader(r)
+	for {
+		header, err := tr.Next()
+		if err == io.EOF {
+			break
+		}
+		AssertNil(t, err)
+
+		if header.Name == path {
+			return true
+		}
+	}
+
+	return false
+}
+
+func AssertBuildpacksHaveDescriptors(t *testing.T, bps []dist.Buildpack, descriptors []dist.BuildpackDescriptor) {
+	AssertEq(t, len(bps), len(descriptors))
+	for _, bp := range bps {
+		found := false
+		for _, descriptor := range descriptors {
+			if diff := cmp.Diff(bp.Descriptor(), descriptor); diff == "" {
+				found = true
+				break
+			}
+		}
+		AssertTrue(t, found)
+	}
+}
+
+func ReadPackConfig(t *testing.T) config.Config {
+	path, err := config.DefaultConfigPath()
+	AssertNil(t, err)
+
+	cfg, err := config.Read(path)
+	AssertNil(t, err)
+	return cfg
+}
+
+func AssertGitHeadEq(t *testing.T, path1, path2 string) {
+	r1, err := git.PlainOpen(path1)
+	AssertNil(t, err)
+
+	r2, err := git.PlainOpen(path2)
+	AssertNil(t, err)
+
+	h1, err := r1.Head()
+	AssertNil(t, err)
+
+	h2, err := r2.Head()
+	AssertNil(t, err)
+
+	AssertEq(t, h1.Hash().String(), h2.Hash().String())
 }
