@@ -1,7 +1,6 @@
 package pack
 
 import (
-	"archive/tar"
 	"bytes"
 	"context"
 	"crypto/sha256"
@@ -39,6 +38,9 @@ import (
 	h "github.com/buildpacks/pack/testhelpers"
 )
 
+// 0.7.5 is the first lifecycle version where both creator and the "lifecycle image" are supported.
+const defaultBuilderLifecycleVersion = "0.7.5"
+
 func TestBuild(t *testing.T) {
 	color.Disable(true)
 	defer color.Disable(false)
@@ -53,13 +55,14 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 		fakeLifecycle         *ifakes.FakeLifecycle
 		defaultBuilderStackID string
 		defaultBuilderImage   *fakes.Image
-		builderName           string
+		defaultBuilderName    string
 		fakeDefaultRunImage   *fakes.Image
 		fakeMirror1           *fakes.Image
 		fakeMirror2           *fakes.Image
 		tmpDir                string
 		outBuf                bytes.Buffer
 		logger                logging.Logger
+		fakeLifecycleImage    *fakes.Image
 	)
 	it.Before(func() {
 		var err error
@@ -70,80 +73,10 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 		tmpDir, err = ioutil.TempDir("", "build-test")
 		h.AssertNil(t, err)
 
-		builderName = "example.com/default/builder:tag"
+		defaultBuilderName = "example.com/default/builder:tag"
 		defaultBuilderStackID = "some.stack.id"
-		defaultBuilderImage = ifakes.NewFakeBuilderImage(t,
-			tmpDir,
-			builderName,
-			defaultBuilderStackID,
-			"1234",
-			"5678",
-			builder.Metadata{
-				Buildpacks: []dist.BuildpackInfo{
-					{ID: "buildpack.1.id", Version: "buildpack.1.version"},
-					{ID: "buildpack.2.id", Version: "buildpack.2.version"},
-				},
-				Stack: builder.StackMetadata{
-					RunImage: builder.RunImageMetadata{
-						Image: "default/run",
-						Mirrors: []string{
-							"registry1.example.com/run/mirror",
-							"registry2.example.com/run/mirror",
-						},
-					},
-				},
-				Lifecycle: builder.LifecycleMetadata{
-					LifecycleInfo: builder.LifecycleInfo{
-						Version: &builder.Version{
-							Version: *semver.MustParse("0.3.0"),
-						},
-					},
-					API: builder.LifecycleAPI{
-						BuildpackVersion: api.MustParse("0.3"),
-						PlatformVersion:  api.MustParse("0.2"),
-					},
-				},
-			},
-			dist.BuildpackLayers{
-				"buildpack.1.id": {
-					"buildpack.1.version": {
-						API: api.MustParse("0.3"),
-						Stacks: []dist.Stack{
-							{
-								ID:     defaultBuilderStackID,
-								Mixins: []string{"mixinX", "build:mixinY", "run:mixinZ"},
-							},
-						},
-					},
-				},
-				"buildpack.2.id": {
-					"buildpack.2.version": {
-						API: api.MustParse("0.3"),
-						Stacks: []dist.Stack{
-							{
-								ID:     defaultBuilderStackID,
-								Mixins: []string{"mixinX", "build:mixinY"},
-							},
-						},
-					},
-				},
-			},
-			dist.Order{{
-				Group: []dist.BuildpackRef{{
-					BuildpackInfo: dist.BuildpackInfo{
-						ID:      "buildpack.1.id",
-						Version: "buildpack.1.version",
-					},
-				}},
-			}, {
-				Group: []dist.BuildpackRef{{
-					BuildpackInfo: dist.BuildpackInfo{
-						ID:      "buildpack.2.id",
-						Version: "buildpack.2.version",
-					},
-				}},
-			}},
-		)
+
+		defaultBuilderImage = newFakeBuilderImage(t, tmpDir, defaultBuilderName, defaultBuilderStackID, defaultBuilderLifecycleVersion)
 		h.AssertNil(t, defaultBuilderImage.SetLabel("io.buildpacks.stack.mixins", `["mixinA", "build:mixinB", "mixinX", "build:mixinY"]`))
 		fakeImageFetcher.LocalImages[defaultBuilderImage.Name()] = defaultBuilderImage
 
@@ -161,6 +94,9 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 		h.AssertNil(t, fakeMirror2.SetLabel("io.buildpacks.stack.id", defaultBuilderStackID))
 		h.AssertNil(t, fakeMirror2.SetLabel("io.buildpacks.stack.mixins", `["mixinA", "mixinX", "run:mixinZ"]`))
 		fakeImageFetcher.LocalImages[fakeMirror2.Name()] = fakeMirror2
+
+		fakeLifecycleImage = fakes.NewImage(fmt.Sprintf("%s:%s", lifecycleImageRepo, defaultBuilderLifecycleVersion), "", nil)
+		fakeImageFetcher.LocalImages[fakeLifecycleImage.Name()] = fakeLifecycleImage
 
 		docker, err := client.NewClientWithOpts(client.FromEnv, client.WithVersion("1.38"))
 		h.AssertNil(t, err)
@@ -185,6 +121,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 		fakeMirror1.Cleanup()
 		fakeMirror2.Cleanup()
 		os.RemoveAll(tmpDir)
+		fakeLifecycleImage.Cleanup()
 	})
 
 	when("#Build", func() {
@@ -192,7 +129,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 			it("is required", func() {
 				h.AssertError(t, subject.Build(context.TODO(), BuildOptions{
 					Image:   "",
-					Builder: builderName,
+					Builder: defaultBuilderName,
 				}),
 					"invalid image name ''",
 				)
@@ -201,7 +138,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 			it("must be a valid image reference", func() {
 				h.AssertError(t, subject.Build(context.TODO(), BuildOptions{
 					Image:   "not@valid",
-					Builder: builderName,
+					Builder: defaultBuilderName,
 				}),
 					"invalid image name 'not@valid'",
 				)
@@ -210,7 +147,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 			it("must be a valid tag reference", func() {
 				h.AssertError(t, subject.Build(context.TODO(), BuildOptions{
 					Image:   "registry.com/my/image@sha256:954e1f01e80ce09d0887ff6ea10b13a812cb01932a0781d6b0cc23f743a874fd",
-					Builder: builderName,
+					Builder: defaultBuilderName,
 				}),
 					"invalid image name 'registry.com/my/image@sha256:954e1f01e80ce09d0887ff6ea10b13a812cb01932a0781d6b0cc23f743a874fd'",
 				)
@@ -218,7 +155,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 
 			it("lifecycle receives resolved reference", func() {
 				h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
-					Builder: builderName,
+					Builder: defaultBuilderName,
 					Image:   "example.com/some/repo:tag",
 				}))
 				h.AssertEq(t, fakeLifecycle.Opts.Image.Context().RegistryStr(), "example.com")
@@ -231,7 +168,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 			it("defaults to the current working directory", func() {
 				h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 					Image:   "some/app",
-					Builder: builderName,
+					Builder: defaultBuilderName,
 				}))
 
 				wd, err := os.Getwd()
@@ -250,7 +187,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 				it(fmt.Sprintf("supports %s files", fileDesc), func() {
 					err := subject.Build(context.TODO(), BuildOptions{
 						Image:   "some/app",
-						Builder: builderName,
+						Builder: defaultBuilderName,
 						AppPath: appPath,
 					})
 					h.AssertNil(t, err)
@@ -269,7 +206,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 				it(fmt.Sprintf("does NOT support %s files", fileDesc), func() {
 					err := subject.Build(context.TODO(), BuildOptions{
 						Image:   "some/app",
-						Builder: builderName,
+						Builder: defaultBuilderName,
 						AppPath: appPath,
 					})
 
@@ -280,7 +217,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 			it("resolves the absolute path", func() {
 				h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 					Image:   "some/app",
-					Builder: builderName,
+					Builder: defaultBuilderName,
 					AppPath: filepath.Join("testdata", "some-app"),
 				}))
 				absPath, err := filepath.Abs(filepath.Join("testdata", "some-app"))
@@ -320,7 +257,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 
 					h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 						Image:   "some/app",
-						Builder: builderName,
+						Builder: defaultBuilderName,
 						AppPath: relLink,
 					}))
 
@@ -333,7 +270,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 
 					h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 						Image:   "some/app",
-						Builder: builderName,
+						Builder: defaultBuilderName,
 						AppPath: relLink,
 					}))
 
@@ -352,7 +289,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 
 					h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 						Image:   "some/app",
-						Builder: builderName,
+						Builder: defaultBuilderName,
 						AppPath: symbolicLink,
 					}))
 
@@ -379,7 +316,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 				it.Before(func() {
 					customBuilderImage = ifakes.NewFakeBuilderImage(t,
 						tmpDir,
-						builderName,
+						defaultBuilderName,
 						"some.stack.id",
 						"1234",
 						"5678",
@@ -392,7 +329,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 							Lifecycle: builder.LifecycleMetadata{
 								LifecycleInfo: builder.LifecycleInfo{
 									Version: &builder.Version{
-										Version: *semver.MustParse("0.3.0"),
+										Version: *semver.MustParse("0.7.5"),
 									},
 								},
 								API: builder.LifecycleAPI{
@@ -420,7 +357,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 				it("it uses the provided builder", func() {
 					h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 						Image:   "some/app",
-						Builder: builderName,
+						Builder: defaultBuilderName,
 					}))
 					h.AssertEq(t, fakeLifecycle.Opts.Builder.Name(), customBuilderImage.Name())
 				})
@@ -447,7 +384,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 				it("uses the provided image", func() {
 					h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 						Image:    "some/app",
-						Builder:  builderName,
+						Builder:  defaultBuilderName,
 						RunImage: "custom/run",
 					}))
 					h.AssertEq(t, fakeLifecycle.Opts.RunImage, "custom/run")
@@ -462,7 +399,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 				it("errors", func() {
 					h.AssertError(t, subject.Build(context.TODO(), BuildOptions{
 						Image:    "some/app",
-						Builder:  builderName,
+						Builder:  defaultBuilderName,
 						RunImage: "custom/run",
 					}),
 						"invalid run-image 'custom/run': run-image stack id 'other.stack' does not match builder stack 'some.stack.id'",
@@ -475,7 +412,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 					it("chooses the best mirror from the builder", func() {
 						h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 							Image:   "some/app",
-							Builder: builderName,
+							Builder: defaultBuilderName,
 						}))
 						h.AssertEq(t, fakeLifecycle.Opts.RunImage, "default/run")
 					})
@@ -483,7 +420,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 					it("chooses the best mirror from the builder", func() {
 						h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 							Image:   "registry1.example.com/some/app",
-							Builder: builderName,
+							Builder: defaultBuilderName,
 						}))
 						h.AssertEq(t, fakeLifecycle.Opts.RunImage, "registry1.example.com/run/mirror")
 					})
@@ -491,7 +428,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 					it("chooses the best mirror from the builder", func() {
 						h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 							Image:   "registry2.example.com/some/app",
-							Builder: builderName,
+							Builder: defaultBuilderName,
 						}))
 						h.AssertEq(t, fakeLifecycle.Opts.RunImage, "registry2.example.com/run/mirror")
 					})
@@ -525,7 +462,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 					it("prefers user provided mirrors", func() {
 						h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 							Image:   "some/app",
-							Builder: builderName,
+							Builder: defaultBuilderName,
 							AdditionalMirrors: map[string][]string{
 								"default/run": {"local/mirror", "registry1.example.com/local/mirror"},
 							},
@@ -536,7 +473,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 					it("choose the correct user provided mirror for the registry", func() {
 						h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 							Image:   "registry1.example.com/some/app",
-							Builder: builderName,
+							Builder: defaultBuilderName,
 							AdditionalMirrors: map[string][]string{
 								"default/run": {"local/mirror", "registry1.example.com/local/mirror"},
 							},
@@ -548,7 +485,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 						it("chooses from builder mirrors", func() {
 							h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 								Image:   "registry2.example.com/some/app",
-								Builder: builderName,
+								Builder: defaultBuilderName,
 								AdditionalMirrors: map[string][]string{
 									"default/run": {"local/mirror", "registry1.example.com/local/mirror"},
 								},
@@ -564,7 +501,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 			it("passes it through to lifecycle", func() {
 				h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 					Image:      "some/app",
-					Builder:    builderName,
+					Builder:    defaultBuilderName,
 					ClearCache: true,
 				}))
 				h.AssertEq(t, fakeLifecycle.Opts.ClearCache, true)
@@ -573,7 +510,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 			it("defaults to false", func() {
 				h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 					Image:   "some/app",
-					Builder: builderName,
+					Builder: defaultBuilderName,
 				}))
 				h.AssertEq(t, fakeLifecycle.Opts.ClearCache, false)
 			})
@@ -589,7 +526,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 			}
 
 			it("builder order is overwritten", func() {
-				additionalBP := createBuildpackTar(t, tmpDir, dist.BuildpackDescriptor{
+				additionalBP := ifakes.CreateBuildpackTar(t, tmpDir, dist.BuildpackDescriptor{
 					API: api.MustParse("0.3"),
 					Info: dist.BuildpackInfo{
 						ID:      "buildpack.add.1.id",
@@ -601,7 +538,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 
 				h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 					Image:      "some/app",
-					Builder:    builderName,
+					Builder:    defaultBuilderName,
 					ClearCache: true,
 					Buildpacks: []string{additionalBP},
 				}))
@@ -619,7 +556,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 				it("resolves version", func() {
 					h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 						Image:      "some/app",
-						Builder:    builderName,
+						Builder:    defaultBuilderName,
 						ClearCache: true,
 						Buildpacks: []string{"buildpack.1.id"},
 					}))
@@ -638,7 +575,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 				it("builder order is prepended", func() {
 					h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 						Image:      "some/app",
-						Builder:    builderName,
+						Builder:    defaultBuilderName,
 						ClearCache: true,
 						Buildpacks: []string{
 							"from=builder:buildpack.1.id@buildpack.1.version",
@@ -657,7 +594,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 
 			when("from=builder is set first", func() {
 				it("builder order is prepended", func() {
-					additionalBP1 := createBuildpackTar(t, tmpDir, dist.BuildpackDescriptor{
+					additionalBP1 := ifakes.CreateBuildpackTar(t, tmpDir, dist.BuildpackDescriptor{
 						API: api.MustParse("0.3"),
 						Info: dist.BuildpackInfo{
 							ID:      "buildpack.add.1.id",
@@ -667,7 +604,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 						Order:  nil,
 					})
 
-					additionalBP2 := createBuildpackTar(t, tmpDir, dist.BuildpackDescriptor{
+					additionalBP2 := ifakes.CreateBuildpackTar(t, tmpDir, dist.BuildpackDescriptor{
 						API: api.MustParse("0.3"),
 						Info: dist.BuildpackInfo{
 							ID:      "buildpack.add.2.id",
@@ -679,7 +616,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 
 					h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 						Image:      "some/app",
-						Builder:    builderName,
+						Builder:    defaultBuilderName,
 						ClearCache: true,
 						Buildpacks: []string{
 							"from=builder",
@@ -721,7 +658,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 
 			when("from=builder is set in middle", func() {
 				it("builder order is appended", func() {
-					additionalBP1 := createBuildpackTar(t, tmpDir, dist.BuildpackDescriptor{
+					additionalBP1 := ifakes.CreateBuildpackTar(t, tmpDir, dist.BuildpackDescriptor{
 						API: api.MustParse("0.3"),
 						Info: dist.BuildpackInfo{
 							ID:      "buildpack.add.1.id",
@@ -731,7 +668,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 						Order:  nil,
 					})
 
-					additionalBP2 := createBuildpackTar(t, tmpDir, dist.BuildpackDescriptor{
+					additionalBP2 := ifakes.CreateBuildpackTar(t, tmpDir, dist.BuildpackDescriptor{
 						API: api.MustParse("0.3"),
 						Info: dist.BuildpackInfo{
 							ID:      "buildpack.add.2.id",
@@ -743,7 +680,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 
 					h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 						Image:      "some/app",
-						Builder:    builderName,
+						Builder:    defaultBuilderName,
 						ClearCache: true,
 						Buildpacks: []string{
 							additionalBP1,
@@ -786,7 +723,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 
 			when("from=builder is set last", func() {
 				it("builder order is appended", func() {
-					additionalBP1 := createBuildpackTar(t, tmpDir, dist.BuildpackDescriptor{
+					additionalBP1 := ifakes.CreateBuildpackTar(t, tmpDir, dist.BuildpackDescriptor{
 						API: api.MustParse("0.3"),
 						Info: dist.BuildpackInfo{
 							ID:      "buildpack.add.1.id",
@@ -796,7 +733,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 						Order:  nil,
 					})
 
-					additionalBP2 := createBuildpackTar(t, tmpDir, dist.BuildpackDescriptor{
+					additionalBP2 := ifakes.CreateBuildpackTar(t, tmpDir, dist.BuildpackDescriptor{
 						API: api.MustParse("0.3"),
 						Info: dist.BuildpackInfo{
 							ID:      "buildpack.add.2.id",
@@ -808,7 +745,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 
 					h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 						Image:      "some/app",
-						Builder:    builderName,
+						Builder:    defaultBuilderName,
 						ClearCache: true,
 						Buildpacks: []string{
 							additionalBP1,
@@ -851,7 +788,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 
 			when("meta-buildpack is used", func() {
 				it("resolves buildpack from builder", func() {
-					buildpackTar := createBuildpackTar(t, tmpDir, dist.BuildpackDescriptor{
+					buildpackTar := ifakes.CreateBuildpackTar(t, tmpDir, dist.BuildpackDescriptor{
 						API: api.MustParse("0.3"),
 						Info: dist.BuildpackInfo{
 							ID:      "metabuildpack.id",
@@ -877,7 +814,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 
 					err := subject.Build(context.TODO(), BuildOptions{
 						Image:      "some/app",
-						Builder:    builderName,
+						Builder:    defaultBuilderName,
 						ClearCache: true,
 						Buildpacks: []string{buildpackTar},
 					})
@@ -890,11 +827,12 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 				var fakePackage *fakes.Image
 
 				it.Before(func() {
-					metaBuildpackTar := createBuildpackTar(t, tmpDir, dist.BuildpackDescriptor{
+					metaBuildpackTar := ifakes.CreateBuildpackTar(t, tmpDir, dist.BuildpackDescriptor{
 						API: api.MustParse("0.3"),
 						Info: dist.BuildpackInfo{
-							ID:      "meta.buildpack.id",
-							Version: "meta.buildpack.version",
+							ID:       "meta.buildpack.id",
+							Version:  "meta.buildpack.version",
+							Homepage: "http://meta.buildpack",
 						},
 						Stacks: nil,
 						Order: dist.Order{{
@@ -908,11 +846,12 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 						}},
 					})
 
-					childBuildpackTar := createBuildpackTar(t, tmpDir, dist.BuildpackDescriptor{
+					childBuildpackTar := ifakes.CreateBuildpackTar(t, tmpDir, dist.BuildpackDescriptor{
 						API: api.MustParse("0.3"),
 						Info: dist.BuildpackInfo{
-							ID:      "child.buildpack.id",
-							Version: "child.buildpack.version",
+							ID:       "child.buildpack.id",
+							Version:  "child.buildpack.version",
+							Homepage: "http://child.buildpack",
 						},
 						Stacks: []dist.Stack{
 							{ID: defaultBuilderStackID},
@@ -969,7 +908,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 				it("all buildpacks are added to ephemeral builder", func() {
 					err := subject.Build(context.TODO(), BuildOptions{
 						Image:      "some/app",
-						Builder:    builderName,
+						Builder:    defaultBuilderName,
 						ClearCache: true,
 						Buildpacks: []string{
 							"example.com/some/package",
@@ -1011,14 +950,14 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 
 					err := subject.Build(context.TODO(), BuildOptions{
 						Image:      "some/app",
-						Builder:    builderName,
+						Builder:    defaultBuilderName,
 						ClearCache: true,
 						Buildpacks: []string{
 							"example.com/some/package",
 						},
 					})
 
-					h.AssertError(t, err, "could not find label 'io.buildpacks.buildpackage.metadata' on image 'example.com/some/package'")
+					h.AssertError(t, err, "extracting buildpacks from 'example.com/some/package': could not find label 'io.buildpacks.buildpackage.metadata'")
 				})
 
 				it("fails when no bp layers label is on package", func() {
@@ -1026,21 +965,21 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 
 					err := subject.Build(context.TODO(), BuildOptions{
 						Image:      "some/app",
-						Builder:    builderName,
+						Builder:    defaultBuilderName,
 						ClearCache: true,
 						Buildpacks: []string{
 							"example.com/some/package",
 						},
 					})
 
-					h.AssertError(t, err, "could not find label 'io.buildpacks.buildpack.layers' on image 'example.com/some/package'")
+					h.AssertError(t, err, "extracting buildpacks from 'example.com/some/package': could not find label 'io.buildpacks.buildpack.layers'")
 				})
 			})
 
 			it("ensures buildpacks exist on builder", func() {
 				h.AssertError(t, subject.Build(context.TODO(), BuildOptions{
 					Image:      "some/app",
-					Builder:    builderName,
+					Builder:    defaultBuilderName,
 					ClearCache: true,
 					Buildpacks: []string{"missing.bp@version"},
 				}),
@@ -1067,7 +1006,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 					it("disallows directory-based buildpacks", func() {
 						err := subject.Build(context.TODO(), BuildOptions{
 							Image:      "some/app",
-							Builder:    builderName,
+							Builder:    defaultBuilderName,
 							ClearCache: true,
 							Buildpacks: []string{
 								"buildpack.1.id@buildpack.1.version",
@@ -1081,7 +1020,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 					it("buildpacks are added to ephemeral builder", func() {
 						err := subject.Build(context.TODO(), BuildOptions{
 							Image:      "some/app",
-							Builder:    builderName,
+							Builder:    defaultBuilderName,
 							ClearCache: true,
 							Buildpacks: []string{
 								"buildpack.1.id@buildpack.1.version",
@@ -1124,7 +1063,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 					it("buildpacks are added to ephemeral builder", func() {
 						err := subject.Build(context.TODO(), BuildOptions{
 							Image:      "some/app",
-							Builder:    builderName,
+							Builder:    defaultBuilderName,
 							ClearCache: true,
 							Buildpacks: []string{
 								"buildpack.1.id@buildpack.1.version",
@@ -1140,7 +1079,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 						h.AssertNil(t, err)
 						buildpack1Info := dist.BuildpackInfo{ID: "buildpack.1.id", Version: "buildpack.1.version"}
 						buildpack2Info := dist.BuildpackInfo{ID: "buildpack.2.id", Version: "buildpack.2.version"}
-						dirBuildpackInfo := dist.BuildpackInfo{ID: "bp.one", Version: "1.2.3"}
+						dirBuildpackInfo := dist.BuildpackInfo{ID: "bp.one", Version: "1.2.3", Homepage: "http://one.buildpack"}
 						tgzBuildpackInfo := dist.BuildpackInfo{ID: "some-other-buildpack-id", Version: "some-other-buildpack-version"}
 						h.AssertEq(t, bldr.Order(), dist.Order{
 							{Group: []dist.BuildpackRef{
@@ -1176,7 +1115,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 					it("adds the buildpack", func() {
 						err := subject.Build(context.TODO(), BuildOptions{
 							Image:      "some/app",
-							Builder:    builderName,
+							Builder:    defaultBuilderName,
 							ClearCache: true,
 							Buildpacks: []string{
 								"buildpack.1.id@buildpack.1.version",
@@ -1213,13 +1152,109 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 					it("returns an error", func() {
 						err := subject.Build(context.TODO(), BuildOptions{
 							Image:   "some/app",
-							Builder: builderName,
+							Builder: defaultBuilderName,
 							Buildpacks: []string{
 								buildpackTgz, // requires mixinA, build:mixinB, run:mixinC
 							},
 						})
 
 						h.AssertError(t, err, "validating stack mixins: buildpack 'some-other-buildpack-id@some-other-buildpack-version' requires missing mixin(s): build:mixinB, mixinA, run:mixinC")
+					})
+				})
+
+				when("buildpack is from a registry", func() {
+					var (
+						fakePackage     *fakes.Image
+						tmpDir          string
+						registryFixture string
+						packHome        string
+					)
+
+					it.Before(func() {
+						var err error
+						tmpDir, err = ioutil.TempDir("", "registry")
+						h.AssertNil(t, err)
+
+						packHome = filepath.Join(tmpDir, ".pack")
+						err = os.MkdirAll(packHome, 0755)
+						h.AssertNil(t, err)
+						os.Setenv("PACK_HOME", packHome)
+
+						registryFixture = h.CreateRegistryFixture(t, tmpDir, filepath.Join("testdata", "registry"))
+
+						childBuildpackTar := ifakes.CreateBuildpackTar(t, tmpDir, dist.BuildpackDescriptor{
+							API: api.MustParse("0.3"),
+							Info: dist.BuildpackInfo{
+								ID:      "example/foo",
+								Version: "1.0.0",
+							},
+							Stacks: []dist.Stack{
+								{ID: defaultBuilderStackID},
+							},
+						})
+
+						bpLayers := dist.BuildpackLayers{
+							"example/foo": {
+								"1.0.0": {
+									API: api.MustParse("0.3"),
+									Stacks: []dist.Stack{
+										{ID: defaultBuilderStackID},
+									},
+									LayerDiffID: diffIDForFile(t, childBuildpackTar),
+								},
+							},
+						}
+
+						md := buildpackage.Metadata{
+							BuildpackInfo: dist.BuildpackInfo{
+								ID:      "example/foo",
+								Version: "1.0.0",
+							},
+							Stacks: []dist.Stack{
+								{ID: defaultBuilderStackID},
+							},
+						}
+
+						fakePackage = fakes.NewImage("example.com/some/package@sha256:8c27fe111c11b722081701dfed3bd55e039b9ce92865473cf4cdfa918071c566", "", nil)
+						h.AssertNil(t, dist.SetLabel(fakePackage, "io.buildpacks.buildpack.layers", bpLayers))
+						h.AssertNil(t, dist.SetLabel(fakePackage, "io.buildpacks.buildpackage.metadata", md))
+
+						h.AssertNil(t, fakePackage.AddLayer(childBuildpackTar))
+
+						fakeImageFetcher.LocalImages[fakePackage.Name()] = fakePackage
+					})
+
+					it.After(func() {
+						os.Unsetenv("PACK_HOME")
+						err := os.RemoveAll(tmpDir)
+						h.AssertNil(t, err)
+					})
+
+					it("all buildpacks are added to ephemeral builder", func() {
+						err := subject.Build(context.TODO(), BuildOptions{
+							Image:      "some/app",
+							Builder:    defaultBuilderName,
+							ClearCache: true,
+							Buildpacks: []string{
+								"urn:cnb:registry:example/foo@1.0.0",
+							},
+							Registry: registryFixture,
+						})
+
+						h.AssertNil(t, err)
+						h.AssertEq(t, fakeLifecycle.Opts.Builder.Name(), defaultBuilderImage.Name())
+						bldr, err := builder.FromImage(defaultBuilderImage)
+						h.AssertNil(t, err)
+						h.AssertEq(t, bldr.Order(), dist.Order{
+							{Group: []dist.BuildpackRef{
+								{BuildpackInfo: dist.BuildpackInfo{ID: "example/foo", Version: "1.0.0"}},
+							}},
+						})
+						h.AssertEq(t, bldr.Buildpacks(), []dist.BuildpackInfo{
+							{ID: "buildpack.1.id", Version: "buildpack.1.version"},
+							{ID: "buildpack.2.id", Version: "buildpack.2.version"},
+							{ID: "example/foo", Version: "1.0.0"},
+						})
 					})
 				})
 			})
@@ -1229,7 +1264,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 			it("should set the env on the ephemeral builder", func() {
 				h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 					Image:   "some/app",
-					Builder: builderName,
+					Builder: defaultBuilderName,
 					Env: map[string]string{
 						"key1": "value1",
 						"key2": "value2",
@@ -1237,30 +1272,41 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 				}))
 				layerTar, err := defaultBuilderImage.FindLayerWithPath("/platform/env/key1")
 				h.AssertNil(t, err)
-				assertTarFileContents(t, layerTar, "/platform/env/key1", `value1`)
-				assertTarFileContents(t, layerTar, "/platform/env/key2", `value2`)
+				h.AssertTarFileContents(t, layerTar, "/platform/env/key1", `value1`)
+				h.AssertTarFileContents(t, layerTar, "/platform/env/key2", `value2`)
 			})
 		})
 
 		when("Publish option", func() {
+			var remoteRunImage, builderWithoutLifecycleImageOrCreator *fakes.Image
+
+			it.Before(func() {
+				remoteRunImage = fakes.NewImage("default/run", "", nil)
+				h.AssertNil(t, remoteRunImage.SetLabel("io.buildpacks.stack.id", defaultBuilderStackID))
+				h.AssertNil(t, remoteRunImage.SetLabel("io.buildpacks.stack.mixins", `["mixinA", "mixinX", "run:mixinZ"]`))
+				fakeImageFetcher.RemoteImages[remoteRunImage.Name()] = remoteRunImage
+
+				builderWithoutLifecycleImageOrCreator = newFakeBuilderImage(
+					t,
+					tmpDir,
+					"example.com/supportscreator/builder:tag",
+					"some.stack.id",
+					"0.3.0",
+				)
+				h.AssertNil(t, builderWithoutLifecycleImageOrCreator.SetLabel("io.buildpacks.stack.mixins", `["mixinA", "build:mixinB", "mixinX", "build:mixinY"]`))
+				fakeImageFetcher.LocalImages[builderWithoutLifecycleImageOrCreator.Name()] = builderWithoutLifecycleImageOrCreator
+			})
+
+			it.After(func() {
+				remoteRunImage.Cleanup()
+				builderWithoutLifecycleImageOrCreator.Cleanup()
+			})
+
 			when("true", func() {
-				var remoteRunImage *fakes.Image
-
-				it.Before(func() {
-					remoteRunImage = fakes.NewImage("default/run", "", nil)
-					h.AssertNil(t, remoteRunImage.SetLabel("io.buildpacks.stack.id", defaultBuilderStackID))
-					h.AssertNil(t, remoteRunImage.SetLabel("io.buildpacks.stack.mixins", `["mixinA", "mixinX", "run:mixinZ"]`))
-					fakeImageFetcher.RemoteImages[remoteRunImage.Name()] = remoteRunImage
-				})
-
-				it.After(func() {
-					remoteRunImage.Cleanup()
-				})
-
 				it("uses a remote run image", func() {
 					h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 						Image:   "some/app",
-						Builder: builderName,
+						Builder: defaultBuilderName,
 						Publish: true,
 					}))
 					h.AssertEq(t, fakeLifecycle.Opts.Publish, true)
@@ -1268,140 +1314,272 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 					args := fakeImageFetcher.FetchCalls["default/run"]
 					h.AssertEq(t, args.Daemon, false)
 
-					args = fakeImageFetcher.FetchCalls[builderName]
+					args = fakeImageFetcher.FetchCalls[defaultBuilderName]
 					h.AssertEq(t, args.Daemon, true)
 				})
 
-				when("false", func() {
-					it("uses a local run image", func() {
-						h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
-							Image:   "some/app",
-							Builder: builderName,
-							Publish: false,
-						}))
-						h.AssertEq(t, fakeLifecycle.Opts.Publish, false)
-
-						args := fakeImageFetcher.FetchCalls["default/run"]
-						h.AssertEq(t, args.Daemon, true)
-						h.AssertEq(t, args.Pull, true)
-
-						args = fakeImageFetcher.FetchCalls[builderName]
-						h.AssertEq(t, args.Daemon, true)
-						h.AssertEq(t, args.Pull, true)
-					})
-				})
-			})
-
-			when("NoPull option", func() {
-				when("true", func() {
-					it("uses the local builder and run images without updating", func() {
-						h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
-							Image:   "some/app",
-							Builder: builderName,
-							NoPull:  true,
-						}))
-
-						args := fakeImageFetcher.FetchCalls["default/run"]
-						h.AssertEq(t, args.Daemon, true)
-						h.AssertEq(t, args.Pull, false)
-
-						args = fakeImageFetcher.FetchCalls[builderName]
-						h.AssertEq(t, args.Daemon, true)
-						h.AssertEq(t, args.Pull, false)
-					})
-				})
-
-				when("false", func() {
-					it("uses pulls the builder and run image before using them", func() {
-						h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
-							Image:   "some/app",
-							Builder: builderName,
-							NoPull:  false,
-						}))
-
-						args := fakeImageFetcher.FetchCalls["default/run"]
-						h.AssertEq(t, args.Daemon, true)
-						h.AssertEq(t, args.Pull, true)
-
-						args = fakeImageFetcher.FetchCalls[builderName]
-						h.AssertEq(t, args.Daemon, true)
-						h.AssertEq(t, args.Pull, true)
-					})
-				})
-			})
-
-			when("ProxyConfig option", func() {
-				when("ProxyConfig is nil", func() {
-					it.Before(func() {
-						h.AssertNil(t, os.Setenv("http_proxy", "other-http-proxy"))
-						h.AssertNil(t, os.Setenv("https_proxy", "other-https-proxy"))
-						h.AssertNil(t, os.Setenv("no_proxy", "other-no-proxy"))
-					})
-
-					when("*_PROXY env vars are set", func() {
-						it.Before(func() {
-							h.AssertNil(t, os.Setenv("HTTP_PROXY", "some-http-proxy"))
-							h.AssertNil(t, os.Setenv("HTTPS_PROXY", "some-https-proxy"))
-							h.AssertNil(t, os.Setenv("NO_PROXY", "some-no-proxy"))
-						})
-
-						it.After(func() {
-							h.AssertNil(t, os.Unsetenv("HTTP_PROXY"))
-							h.AssertNil(t, os.Unsetenv("HTTPS_PROXY"))
-							h.AssertNil(t, os.Unsetenv("NO_PROXY"))
-						})
-
-						it("defaults to the *_PROXY environment variables", func() {
+				when("builder is untrusted", func() {
+					when("lifecycle image is available", func() {
+						it("uses the 5 phases with the lifecycle image", func() {
 							h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
-								Image:   "some/app",
-								Builder: builderName,
+								Image:        "some/app",
+								Builder:      defaultBuilderName,
+								Publish:      true,
+								TrustBuilder: false,
 							}))
-							h.AssertEq(t, fakeLifecycle.Opts.HTTPProxy, "some-http-proxy")
-							h.AssertEq(t, fakeLifecycle.Opts.HTTPSProxy, "some-https-proxy")
-							h.AssertEq(t, fakeLifecycle.Opts.NoProxy, "some-no-proxy")
+							h.AssertEq(t, fakeLifecycle.Opts.UseCreator, false)
+							h.AssertEq(t, fakeLifecycle.Opts.LifecycleImage, fakeLifecycleImage.Name())
+
+							args := fakeImageFetcher.FetchCalls[fakeLifecycleImage.Name()]
+							h.AssertEq(t, args.Daemon, true)
+							h.AssertEq(t, args.Pull, true)
 						})
 					})
 
-					it("falls back to the *_proxy environment variables", func() {
-						h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
-							Image:   "some/app",
-							Builder: builderName,
-						}))
-						h.AssertEq(t, fakeLifecycle.Opts.HTTPProxy, "other-http-proxy")
-						h.AssertEq(t, fakeLifecycle.Opts.HTTPSProxy, "other-https-proxy")
-						h.AssertEq(t, fakeLifecycle.Opts.NoProxy, "other-no-proxy")
+					when("lifecycle image is not available", func() {
+						it("errors", func() {
+							h.AssertNotNil(t, subject.Build(context.TODO(), BuildOptions{
+								Image:        "some/app",
+								Builder:      builderWithoutLifecycleImageOrCreator.Name(),
+								Publish:      true,
+								TrustBuilder: false,
+							}))
+						})
 					})
-				}, spec.Sequential())
+				})
 
-				when("ProxyConfig is not nil", func() {
-					it("passes the values through", func() {
-						h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
-							Image:   "some/app",
-							Builder: builderName,
-							ProxyConfig: &ProxyConfig{
-								HTTPProxy:  "custom-http-proxy",
-								HTTPSProxy: "custom-https-proxy",
-								NoProxy:    "custom-no-proxy",
-							},
-						}))
-						h.AssertEq(t, fakeLifecycle.Opts.HTTPProxy, "custom-http-proxy")
-						h.AssertEq(t, fakeLifecycle.Opts.HTTPSProxy, "custom-https-proxy")
-						h.AssertEq(t, fakeLifecycle.Opts.NoProxy, "custom-no-proxy")
+				when("builder is trusted", func() {
+					when("lifecycle supports creator", func() {
+						it("uses the creator with the provided builder", func() {
+							h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
+								Image:        "some/app",
+								Builder:      defaultBuilderName,
+								Publish:      true,
+								TrustBuilder: true,
+							}))
+							h.AssertEq(t, fakeLifecycle.Opts.UseCreator, true)
+
+							args := fakeImageFetcher.FetchCalls[fakeLifecycleImage.Name()]
+							h.AssertNil(t, args)
+						})
+					})
+
+					when("lifecycle doesn't support creator", func() {
+						// the default test builder (example.com/default/builder:tag) has lifecycle version 0.3.0, so creator is not supported
+						it("uses the 5 phases with the provided builder", func() {
+							h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
+								Image:        "some/app",
+								Builder:      builderWithoutLifecycleImageOrCreator.Name(),
+								Publish:      true,
+								TrustBuilder: true,
+							}))
+							h.AssertEq(t, fakeLifecycle.Opts.UseCreator, false)
+							h.AssertEq(t, fakeLifecycle.Opts.LifecycleImage, builderWithoutLifecycleImageOrCreator.Name())
+
+							args := fakeImageFetcher.FetchCalls[fakeLifecycleImage.Name()]
+							h.AssertNil(t, args)
+						})
 					})
 				})
 			})
 
-			when("Network option", func() {
-				it("passes the value through", func() {
+			when("false", func() {
+				it("uses a local run image", func() {
 					h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
 						Image:   "some/app",
-						Builder: builderName,
-						ContainerConfig: ContainerConfig{
-							Network: "some-network",
+						Builder: defaultBuilderName,
+						Publish: false,
+					}))
+					h.AssertEq(t, fakeLifecycle.Opts.Publish, false)
+
+					args := fakeImageFetcher.FetchCalls["default/run"]
+					h.AssertEq(t, args.Daemon, true)
+					h.AssertEq(t, args.Pull, true)
+
+					args = fakeImageFetcher.FetchCalls[defaultBuilderName]
+					h.AssertEq(t, args.Daemon, true)
+					h.AssertEq(t, args.Pull, true)
+				})
+
+				when("builder is untrusted", func() {
+					when("lifecycle image is available", func() {
+						it("uses the 5 phases with the lifecycle image", func() {
+							h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
+								Image:        "some/app",
+								Builder:      defaultBuilderName,
+								Publish:      false,
+								TrustBuilder: false,
+							}))
+							h.AssertEq(t, fakeLifecycle.Opts.UseCreator, false)
+							h.AssertEq(t, fakeLifecycle.Opts.LifecycleImage, fakeLifecycleImage.Name())
+
+							args := fakeImageFetcher.FetchCalls[fakeLifecycleImage.Name()]
+							h.AssertEq(t, args.Daemon, true)
+							h.AssertEq(t, args.Pull, true)
+						})
+					})
+
+					when("lifecycle image is not available", func() {
+						it("errors", func() {
+							h.AssertNotNil(t, subject.Build(context.TODO(), BuildOptions{
+								Image:        "some/app",
+								Builder:      builderWithoutLifecycleImageOrCreator.Name(),
+								Publish:      false,
+								TrustBuilder: false,
+							}))
+						})
+					})
+				})
+
+				when("builder is trusted", func() {
+					when("lifecycle supports creator", func() {
+						it("uses the creator with the provided builder", func() {
+							h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
+								Image:        "some/app",
+								Builder:      defaultBuilderName,
+								Publish:      false,
+								TrustBuilder: true,
+							}))
+							h.AssertEq(t, fakeLifecycle.Opts.UseCreator, true)
+
+							args := fakeImageFetcher.FetchCalls[fakeLifecycleImage.Name()]
+							h.AssertNil(t, args)
+						})
+					})
+
+					when("lifecycle doesn't support creator", func() {
+						// the default test builder (example.com/default/builder:tag) has lifecycle version 0.3.0, so creator is not supported
+						it("uses the 5 phases with the provided builder", func() {
+							h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
+								Image:        "some/app",
+								Builder:      builderWithoutLifecycleImageOrCreator.Name(),
+								Publish:      false,
+								TrustBuilder: true,
+							}))
+							h.AssertEq(t, fakeLifecycle.Opts.UseCreator, false)
+							h.AssertEq(t, fakeLifecycle.Opts.LifecycleImage, builderWithoutLifecycleImageOrCreator.Name())
+
+							args := fakeImageFetcher.FetchCalls[fakeLifecycleImage.Name()]
+							h.AssertNil(t, args)
+						})
+					})
+				})
+			})
+		})
+
+		when("NoPull option", func() {
+			when("true", func() {
+				it("uses the local builder and run images without updating", func() {
+					h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
+						Image:   "some/app",
+						Builder: defaultBuilderName,
+						NoPull:  true,
+					}))
+
+					args := fakeImageFetcher.FetchCalls["default/run"]
+					h.AssertEq(t, args.Daemon, true)
+					h.AssertEq(t, args.Pull, false)
+
+					args = fakeImageFetcher.FetchCalls[defaultBuilderName]
+					h.AssertEq(t, args.Daemon, true)
+					h.AssertEq(t, args.Pull, false)
+
+					args = fakeImageFetcher.FetchCalls["buildpacksio/lifecycle:0.7.5"]
+					h.AssertEq(t, args.Daemon, true)
+					h.AssertEq(t, args.Pull, false)
+				})
+			})
+
+			when("false", func() {
+				it("uses pulls the builder and run image before using them", func() {
+					h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
+						Image:   "some/app",
+						Builder: defaultBuilderName,
+						NoPull:  false,
+					}))
+
+					args := fakeImageFetcher.FetchCalls["default/run"]
+					h.AssertEq(t, args.Daemon, true)
+					h.AssertEq(t, args.Pull, true)
+
+					args = fakeImageFetcher.FetchCalls[defaultBuilderName]
+					h.AssertEq(t, args.Daemon, true)
+					h.AssertEq(t, args.Pull, true)
+				})
+			})
+		})
+
+		when("ProxyConfig option", func() {
+			when("ProxyConfig is nil", func() {
+				it.Before(func() {
+					h.AssertNil(t, os.Setenv("http_proxy", "other-http-proxy"))
+					h.AssertNil(t, os.Setenv("https_proxy", "other-https-proxy"))
+					h.AssertNil(t, os.Setenv("no_proxy", "other-no-proxy"))
+				})
+
+				when("*_PROXY env vars are set", func() {
+					it.Before(func() {
+						h.AssertNil(t, os.Setenv("HTTP_PROXY", "some-http-proxy"))
+						h.AssertNil(t, os.Setenv("HTTPS_PROXY", "some-https-proxy"))
+						h.AssertNil(t, os.Setenv("NO_PROXY", "some-no-proxy"))
+					})
+
+					it.After(func() {
+						h.AssertNil(t, os.Unsetenv("HTTP_PROXY"))
+						h.AssertNil(t, os.Unsetenv("HTTPS_PROXY"))
+						h.AssertNil(t, os.Unsetenv("NO_PROXY"))
+					})
+
+					it("defaults to the *_PROXY environment variables", func() {
+						h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
+							Image:   "some/app",
+							Builder: defaultBuilderName,
+						}))
+						h.AssertEq(t, fakeLifecycle.Opts.HTTPProxy, "some-http-proxy")
+						h.AssertEq(t, fakeLifecycle.Opts.HTTPSProxy, "some-https-proxy")
+						h.AssertEq(t, fakeLifecycle.Opts.NoProxy, "some-no-proxy")
+					})
+				})
+
+				it("falls back to the *_proxy environment variables", func() {
+					h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
+						Image:   "some/app",
+						Builder: defaultBuilderName,
+					}))
+					h.AssertEq(t, fakeLifecycle.Opts.HTTPProxy, "other-http-proxy")
+					h.AssertEq(t, fakeLifecycle.Opts.HTTPSProxy, "other-https-proxy")
+					h.AssertEq(t, fakeLifecycle.Opts.NoProxy, "other-no-proxy")
+				})
+			}, spec.Sequential())
+
+			when("ProxyConfig is not nil", func() {
+				it("passes the values through", func() {
+					h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
+						Image:   "some/app",
+						Builder: defaultBuilderName,
+						ProxyConfig: &ProxyConfig{
+							HTTPProxy:  "custom-http-proxy",
+							HTTPSProxy: "custom-https-proxy",
+							NoProxy:    "custom-no-proxy",
 						},
 					}))
-					h.AssertEq(t, fakeLifecycle.Opts.Network, "some-network")
+					h.AssertEq(t, fakeLifecycle.Opts.HTTPProxy, "custom-http-proxy")
+					h.AssertEq(t, fakeLifecycle.Opts.HTTPSProxy, "custom-https-proxy")
+					h.AssertEq(t, fakeLifecycle.Opts.NoProxy, "custom-no-proxy")
 				})
+			})
+		})
+
+		when("Network option", func() {
+			it("passes the value through", func() {
+				h.AssertNil(t, subject.Build(context.TODO(), BuildOptions{
+					Image:   "some/app",
+					Builder: defaultBuilderName,
+					ContainerConfig: ContainerConfig{
+						Network: "some-network",
+					},
+				}))
+				h.AssertEq(t, fakeLifecycle.Opts.Network, "some-network")
 			})
 		})
 
@@ -1417,7 +1595,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 						it.Before(func() {
 							compatibleBuilder = ifakes.NewFakeBuilderImage(t,
 								tmpDir,
-								"compatible-"+builderName,
+								"compatible-"+defaultBuilderName,
 								defaultBuilderStackID,
 								"1234",
 								"5678",
@@ -1434,7 +1612,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 									Lifecycle: builder.LifecycleMetadata{
 										LifecycleInfo: builder.LifecycleInfo{
 											Version: &builder.Version{
-												Version: *semver.MustParse("0.3.0"),
+												Version: *semver.MustParse("0.7.5"),
 											},
 										},
 										API: builder.LifecycleAPI{
@@ -1466,7 +1644,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 					it.Before(func() {
 						incompatibleBuilderImage = ifakes.NewFakeBuilderImage(t,
 							tmpDir,
-							"incompatible-"+builderName,
+							"incompatible-"+defaultBuilderName,
 							defaultBuilderStackID,
 							"1234",
 							"5678",
@@ -1483,7 +1661,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 								Lifecycle: builder.LifecycleMetadata{
 									LifecycleInfo: builder.LifecycleInfo{
 										Version: &builder.Version{
-											Version: *semver.MustParse("0.3.0"),
+											Version: *semver.MustParse("0.7.5"),
 										},
 									},
 									API: builder.LifecycleAPI{
@@ -1527,7 +1705,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 				it("returns an error", func() {
 					err := subject.Build(context.TODO(), BuildOptions{
 						Image:   "some/app",
-						Builder: builderName,
+						Builder: defaultBuilderName,
 					})
 
 					h.AssertError(t, err, "validating stack mixins: 'default/run' missing required mixin(s): mixinA")
@@ -1543,7 +1721,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 				it("returns an error", func() {
 					err := subject.Build(context.TODO(), BuildOptions{
 						Image:   "some/app",
-						Builder: builderName,
+						Builder: defaultBuilderName,
 					})
 
 					h.AssertError(t, err, "validating stack mixins: buildpack 'buildpack.1.id@buildpack.1.version' requires missing mixin(s): build:mixinY, mixinX, run:mixinZ")
@@ -1560,7 +1738,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 				it("prepends /platform to the mount paths", func() {
 					subject.Build(context.TODO(), BuildOptions{
 						Image:   "some/app",
-						Builder: builderName,
+						Builder: defaultBuilderName,
 						ContainerConfig: ContainerConfig{
 							Volumes: []string{"/a:/x", "/b:/some/path/y"},
 						},
@@ -1576,7 +1754,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 					it("returns an error", func() {
 						err := subject.Build(context.TODO(), BuildOptions{
 							Image:   "some/app",
-							Builder: builderName,
+							Builder: defaultBuilderName,
 							ContainerConfig: ContainerConfig{
 								Volumes: []string{"/a:/x", ":::"},
 							},
@@ -1596,7 +1774,7 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 					volume := fmt.Sprintf("%v:/x", dir)
 					err := subject.Build(context.TODO(), BuildOptions{
 						Image:   "some/app",
-						Builder: builderName,
+						Builder: defaultBuilderName,
 						ContainerConfig: ContainerConfig{
 							Volumes: []string{volume},
 						},
@@ -1614,55 +1792,6 @@ func testBuild(t *testing.T, when spec.G, it spec.S) {
 	})
 }
 
-func assertTarFileContents(t *testing.T, tarfile, path, expected string) {
-	t.Helper()
-	exist, contents := tarFileContents(t, tarfile, path)
-	if !exist {
-		t.Fatalf("%s does not exist in %s", path, tarfile)
-	}
-	h.AssertEq(t, contents, expected)
-}
-
-func tarFileContents(t *testing.T, tarfile, path string) (exist bool, contents string) {
-	t.Helper()
-	r, err := os.Open(tarfile)
-	h.AssertNil(t, err)
-	defer r.Close()
-
-	tr := tar.NewReader(r)
-	for {
-		header, err := tr.Next()
-		if err == io.EOF {
-			break
-		}
-		h.AssertNil(t, err)
-
-		if header.Name == path {
-			buf, err := ioutil.ReadAll(tr)
-			h.AssertNil(t, err)
-			return true, string(buf)
-		}
-	}
-	return false, ""
-}
-
-func createBuildpackTar(t *testing.T, tmpDir string, descriptor dist.BuildpackDescriptor) string {
-	buildpack, err := ifakes.NewFakeBuildpackBlob(descriptor, 0777)
-	h.AssertNil(t, err)
-
-	tempFile, err := ioutil.TempFile(tmpDir, "bp-*.tar")
-	h.AssertNil(t, err)
-	defer tempFile.Close()
-
-	reader, err := buildpack.Open()
-	h.AssertNil(t, err)
-
-	_, err = io.Copy(tempFile, reader)
-	h.AssertNil(t, err)
-
-	return tempFile.Name()
-}
-
 func diffIDForFile(t *testing.T, path string) string {
 	file, err := os.Open(path)
 	h.AssertNil(t, err)
@@ -1672,4 +1801,79 @@ func diffIDForFile(t *testing.T, path string) string {
 	h.AssertNil(t, err)
 
 	return "sha256:" + hex.EncodeToString(hasher.Sum(make([]byte, 0, hasher.Size())))
+}
+
+func newFakeBuilderImage(t *testing.T, tmpDir, builderName, defaultBuilderStackID, lifecycleVersion string) *fakes.Image {
+	return ifakes.NewFakeBuilderImage(t,
+		tmpDir,
+		builderName,
+		defaultBuilderStackID,
+		"1234",
+		"5678",
+		builder.Metadata{
+			Buildpacks: []dist.BuildpackInfo{
+				{ID: "buildpack.1.id", Version: "buildpack.1.version"},
+				{ID: "buildpack.2.id", Version: "buildpack.2.version"},
+			},
+			Stack: builder.StackMetadata{
+				RunImage: builder.RunImageMetadata{
+					Image: "default/run",
+					Mirrors: []string{
+						"registry1.example.com/run/mirror",
+						"registry2.example.com/run/mirror",
+					},
+				},
+			},
+			Lifecycle: builder.LifecycleMetadata{
+				LifecycleInfo: builder.LifecycleInfo{
+					Version: &builder.Version{
+						Version: *semver.MustParse(lifecycleVersion),
+					},
+				},
+				API: builder.LifecycleAPI{
+					BuildpackVersion: api.MustParse("0.3"),
+					PlatformVersion:  api.MustParse("0.2"),
+				},
+			},
+		},
+		dist.BuildpackLayers{
+			"buildpack.1.id": {
+				"buildpack.1.version": {
+					API: api.MustParse("0.3"),
+					Stacks: []dist.Stack{
+						{
+							ID:     defaultBuilderStackID,
+							Mixins: []string{"mixinX", "build:mixinY", "run:mixinZ"},
+						},
+					},
+				},
+			},
+			"buildpack.2.id": {
+				"buildpack.2.version": {
+					API: api.MustParse("0.3"),
+					Stacks: []dist.Stack{
+						{
+							ID:     defaultBuilderStackID,
+							Mixins: []string{"mixinX", "build:mixinY"},
+						},
+					},
+				},
+			},
+		},
+		dist.Order{{
+			Group: []dist.BuildpackRef{{
+				BuildpackInfo: dist.BuildpackInfo{
+					ID:      "buildpack.1.id",
+					Version: "buildpack.1.version",
+				},
+			}},
+		}, {
+			Group: []dist.BuildpackRef{{
+				BuildpackInfo: dist.BuildpackInfo{
+					ID:      "buildpack.2.id",
+					Version: "buildpack.2.version",
+				},
+			}},
+		}},
+	)
 }
