@@ -8,25 +8,28 @@ import (
 	"strings"
 
 	"github.com/pkg/errors"
+	ignore "github.com/sabhiram/go-gitignore"
 	"github.com/spf13/cobra"
 
 	"github.com/buildpacks/pack"
 	"github.com/buildpacks/pack/internal/config"
 	"github.com/buildpacks/pack/internal/paths"
-	"github.com/buildpacks/pack/internal/project"
 	"github.com/buildpacks/pack/internal/style"
 	"github.com/buildpacks/pack/logging"
+	"github.com/buildpacks/pack/project"
 )
 
 type BuildFlags struct {
 	AppPath            string
 	Builder            string
+	Registry           string
 	RunImage           string
 	Env                []string
 	EnvFiles           []string
 	Publish            bool
 	NoPull             bool
 	ClearCache         bool
+	TrustBuilder       bool
 	Buildpacks         []string
 	Network            string
 	DescriptorPath     string
@@ -34,6 +37,7 @@ type BuildFlags struct {
 	DefaultProcessType string
 }
 
+// Build an image from source code
 func Build(logger logging.Logger, cfg config.Config, packClient PackClient) *cobra.Command {
 	var flags BuildFlags
 
@@ -42,11 +46,11 @@ func Build(logger logging.Logger, cfg config.Config, packClient PackClient) *cob
 		Args:  cobra.ExactArgs(1),
 		Short: "Generate app image from source code",
 		RunE: logError(logger, func(cmd *cobra.Command, args []string) error {
-			imageName := args[0]
-			if flags.Builder == "" {
-				suggestSettingBuilder(logger, packClient)
-				return MakeSoftError()
+			if err := validateBuildFlags(flags, logger, cfg, packClient); err != nil {
+				return err
 			}
+
+			imageName := args[0]
 
 			descriptor, actualDescriptorPath, err := parseProjectToml(flags.AppPath, flags.DescriptorPath)
 			if err != nil {
@@ -54,6 +58,11 @@ func Build(logger logging.Logger, cfg config.Config, packClient PackClient) *cob
 			}
 			if actualDescriptorPath != "" {
 				logger.Debugf("Using project descriptor located at %s", style.Symbol(actualDescriptorPath))
+			}
+
+			fileFilter, err := getFileFilter(descriptor)
+			if err != nil {
+				return err
 			}
 
 			env, err := parseEnv(descriptor, flags.EnvFiles, flags.Env)
@@ -80,9 +89,27 @@ func Build(logger logging.Logger, cfg config.Config, packClient PackClient) *cob
 				}
 			}
 
+			var cfgTrustedBuilder bool
+			for _, trustedBuilder := range cfg.TrustedBuilders {
+				logger.Debugf("Builder %s is trusted", style.Symbol(trustedBuilder.Name))
+				if flags.Builder == trustedBuilder.Name {
+					cfgTrustedBuilder = true
+					break
+				}
+			}
+
+			var suggestedBuilder bool
+			for _, builder := range suggestedBuilders {
+				if flags.Builder == builder.Image {
+					suggestedBuilder = true
+					break
+				}
+			}
+
 			if err := packClient.Build(cmd.Context(), pack.BuildOptions{
 				AppPath:           flags.AppPath,
 				Builder:           flags.Builder,
+				Registry:          flags.Registry,
 				AdditionalMirrors: getMirrors(cfg),
 				RunImage:          flags.RunImage,
 				Env:               env,
@@ -90,14 +117,16 @@ func Build(logger logging.Logger, cfg config.Config, packClient PackClient) *cob
 				Publish:           flags.Publish,
 				NoPull:            flags.NoPull,
 				ClearCache:        flags.ClearCache,
+				TrustBuilder:      flags.TrustBuilder || cfgTrustedBuilder || suggestedBuilder,
 				Buildpacks:        buildpacks,
 				ContainerConfig: pack.ContainerConfig{
 					Network: flags.Network,
 					Volumes: flags.Volumes,
 				},
 				DefaultProcessType: flags.DefaultProcessType,
+				FileFilter:         fileFilter,
 			}); err != nil {
-				return err
+				return errors.Wrap(err, "failed to build")
 			}
 			logger.Infof("Successfully built image %s", style.Symbol(imageName))
 			return nil
@@ -112,16 +141,34 @@ func Build(logger logging.Logger, cfg config.Config, packClient PackClient) *cob
 func buildCommandFlags(cmd *cobra.Command, buildFlags *BuildFlags, cfg config.Config) {
 	cmd.Flags().StringVarP(&buildFlags.AppPath, "path", "p", "", "Path to app dir or zip-formatted file (defaults to current working directory)")
 	cmd.Flags().StringVarP(&buildFlags.Builder, "builder", "B", cfg.DefaultBuilder, "Builder image")
+	cmd.Flags().StringVarP(&buildFlags.Registry, "buildpack-registry", "R", cfg.DefaultRegistry, "Buildpack Registry URL")
+	if !cfg.Experimental {
+		cmd.Flags().MarkHidden("buildpack-registry")
+	}
 	cmd.Flags().StringVar(&buildFlags.RunImage, "run-image", "", "Run image (defaults to default stack's run image)")
 	cmd.Flags().StringArrayVarP(&buildFlags.Env, "env", "e", []string{}, "Build-time environment variable, in the form 'VAR=VALUE' or 'VAR'.\nWhen using latter value-less form, value will be taken from current\n  environment at the time this command is executed.\nThis flag may be specified multiple times and will override\n  individual values defined by --env-file.")
 	cmd.Flags().StringArrayVar(&buildFlags.EnvFiles, "env-file", []string{}, "Build-time environment variables file\nOne variable per line, of the form 'VAR=VALUE' or 'VAR'\nWhen using latter value-less form, value will be taken from current\n  environment at the time this command is executed")
 	cmd.Flags().BoolVar(&buildFlags.NoPull, "no-pull", false, "Skip pulling builder and run images before use")
 	cmd.Flags().BoolVar(&buildFlags.ClearCache, "clear-cache", false, "Clear image's associated cache before building")
+	cmd.Flags().BoolVar(&buildFlags.TrustBuilder, "trust-builder", false, "Trust the provided builder\nAll lifecycle phases will be run in a single container (if supported by the lifecycle).")
 	cmd.Flags().StringSliceVarP(&buildFlags.Buildpacks, "buildpack", "b", nil, "Buildpack reference in the form of '<buildpack>@<version>',\n  path to a buildpack directory (not supported on Windows),\n  path/URL to a buildpack .tar or .tgz file, or\n  the name of a packaged buildpack image"+multiValueHelp("buildpack"))
 	cmd.Flags().StringVar(&buildFlags.Network, "network", "", "Connect detect and build containers to network")
 	cmd.Flags().StringVarP(&buildFlags.DescriptorPath, "descriptor", "d", "", "Path to the project descriptor file")
 	cmd.Flags().StringArrayVar(&buildFlags.Volumes, "volume", nil, "Mount host volume into the build container, in the form '<host path>:<target path>'. Target path will be prefixed with '/platform/'"+multiValueHelp("volume"))
 	cmd.Flags().StringVarP(&buildFlags.DefaultProcessType, "default-process", "D", "", "Set the default process type")
+}
+
+func validateBuildFlags(flags BuildFlags, logger logging.Logger, cfg config.Config, packClient PackClient) error {
+	if flags.Builder == "" {
+		suggestSettingBuilder(logger, packClient)
+		return pack.NewSoftError()
+	}
+
+	if flags.Registry != "" && !cfg.Experimental {
+		return pack.NewExperimentError("Support for buildpack registries is currently experimental.")
+	}
+
+	return nil
 }
 
 func parseEnv(project project.Descriptor, envFiles []string, envVars []string) (map[string]string, error) {
@@ -189,4 +236,25 @@ func parseProjectToml(appPath, descriptorPath string) (project.Descriptor, strin
 
 	descriptor, err := project.ReadProjectDescriptor(actualPath)
 	return descriptor, actualPath, err
+}
+
+func getFileFilter(descriptor project.Descriptor) (func(string) bool, error) {
+	if len(descriptor.Build.Exclude) > 0 {
+		excludes, err := ignore.CompileIgnoreLines(descriptor.Build.Exclude...)
+		if err != nil {
+			return nil, err
+		}
+		return func(fileName string) bool {
+			return !excludes.MatchesPath(fileName)
+		}, nil
+	}
+	if len(descriptor.Build.Include) > 0 {
+		includes, err := ignore.CompileIgnoreLines(descriptor.Build.Include...)
+		if err != nil {
+			return nil, err
+		}
+		return includes.MatchesPath, nil
+	}
+
+	return nil, nil
 }
