@@ -10,6 +10,7 @@ import (
 	"testing"
 
 	"github.com/golang/mock/gomock"
+	"github.com/pkg/errors"
 	"github.com/sclevine/spec"
 	"github.com/sclevine/spec/report"
 	"github.com/spf13/cobra"
@@ -47,6 +48,18 @@ func testBuildCommand(t *testing.T, when spec.G, it spec.S) {
 	})
 
 	when("#BuildCommand", func() {
+		when("no builder is specified", func() {
+			it("returns a soft error", func() {
+				mockClient.EXPECT().InspectBuilder(gomock.Any(), false).Return(&pack.BuilderInfo{
+					Description: "",
+				}, nil).AnyTimes()
+
+				command.SetArgs([]string{"image"})
+				err := command.Execute()
+				h.AssertError(t, err, pack.NewSoftError().Error())
+			})
+		})
+
 		when("a builder and image are set", func() {
 			it("builds an image with a builder", func() {
 				mockClient.EXPECT().
@@ -63,6 +76,39 @@ func testBuildCommand(t *testing.T, when spec.G, it spec.S) {
 
 				command.SetArgs([]string{"-B", "my-builder", "image"})
 				h.AssertNil(t, command.Execute())
+			})
+			when("the builder is trusted", func() {
+				it("sets the trust builder option", func() {
+					mockClient.EXPECT().
+						Build(gomock.Any(), EqBuildOptionsWithTrustedBuilder(true)).
+						Return(nil)
+
+					cfg := config.Config{TrustedBuilders: []config.TrustedBuilder{{Name: "my-builder"}}}
+					command := commands.Build(logger, cfg, mockClient)
+
+					command.SetArgs([]string{"image", "--builder", "my-builder"})
+					h.AssertNil(t, command.Execute())
+				})
+			})
+
+			when("the builder is suggested", func() {
+				it("sets the trust builder option", func() {
+					mockClient.EXPECT().
+						Build(gomock.Any(), EqBuildOptionsWithTrustedBuilder(true)).
+						Return(nil)
+
+					command.SetArgs([]string{"image", "--builder", "heroku/buildpacks:18"})
+					h.AssertNil(t, command.Execute())
+				})
+			})
+		})
+
+		when("--buildpack-registry flag is specified but experimental isn't set in the config", func() {
+			it("errors with a descriptive message", func() {
+				command.SetArgs([]string{"image", "--builder", "my-builder", "--buildpack-registry", "some-registry"})
+				err := command.Execute()
+				h.AssertNotNil(t, err)
+				h.AssertError(t, err, "Support for buildpack registries is currently experimental.")
 			})
 		})
 
@@ -99,69 +145,144 @@ func testBuildCommand(t *testing.T, when spec.G, it spec.S) {
 			})
 		})
 
-		when("an env file is provided", func() {
-			var envPath string
+		when("env file", func() {
+			when("an env file is provided", func() {
+				var envPath string
+
+				it.Before(func() {
+					envfile, err := ioutil.TempFile("", "envfile")
+					h.AssertNil(t, err)
+					defer envfile.Close()
+
+					envfile.WriteString(`KEY=VALUE`)
+					envPath = envfile.Name()
+				})
+
+				it.After(func() {
+					h.AssertNil(t, os.RemoveAll(envPath))
+				})
+
+				it("builds an image env variables read from the env file", func() {
+					mockClient.EXPECT().
+						Build(gomock.Any(), EqBuildOptionsWithEnv(map[string]string{
+							"KEY": "VALUE",
+						})).
+						Return(nil)
+
+					command.SetArgs([]string{"--builder", "my-builder", "image", "--env-file", envPath})
+					h.AssertNil(t, command.Execute())
+				})
+			})
+
+			when("a env file is provided but doesn't exist", func() {
+				it("fails to run", func() {
+					command.SetArgs([]string{"--builder", "my-builder", "image", "--env-file", ""})
+					err := command.Execute()
+					h.AssertError(t, err, "parse env file")
+				})
+			})
+
+			when("an empty env file is provided", func() {
+				var envPath string
+
+				it.Before(func() {
+					envfile, err := ioutil.TempFile("", "envfile")
+					h.AssertNil(t, err)
+					defer envfile.Close()
+
+					envfile.WriteString(``)
+					envPath = envfile.Name()
+				})
+
+				it.After(func() {
+					h.AssertNil(t, os.RemoveAll(envPath))
+				})
+
+				it("successfully builds", func() {
+					mockClient.EXPECT().
+						Build(gomock.Any(), EqBuildOptionsWithEnv(map[string]string{})).
+						Return(nil)
+
+					command.SetArgs([]string{"--builder", "my-builder", "image", "--env-file", envPath})
+					h.AssertNil(t, command.Execute())
+				})
+			})
+
+			when("two env files are provided with conflicted keys", func() {
+				var envPath1 string
+				var envPath2 string
+
+				it.Before(func() {
+					envfile1, err := ioutil.TempFile("", "envfile")
+					h.AssertNil(t, err)
+					defer envfile1.Close()
+
+					envfile1.WriteString("KEY1=VALUE1\nKEY2=IGNORED")
+					envPath1 = envfile1.Name()
+
+					envfile2, err := ioutil.TempFile("", "envfile")
+					h.AssertNil(t, err)
+					defer envfile2.Close()
+
+					envfile2.WriteString("KEY2=VALUE2")
+					envPath2 = envfile2.Name()
+				})
+
+				it.After(func() {
+					h.AssertNil(t, os.RemoveAll(envPath1))
+					h.AssertNil(t, os.RemoveAll(envPath2))
+				})
+
+				it("builds an image with the last value of each env variable", func() {
+					mockClient.EXPECT().
+						Build(gomock.Any(), EqBuildOptionsWithEnv(map[string]string{
+							"KEY1": "VALUE1",
+							"KEY2": "VALUE2",
+						})).
+						Return(nil)
+
+					command.SetArgs([]string{"--builder", "my-builder", "image", "--env-file", envPath1, "--env-file", envPath2})
+					h.AssertNil(t, command.Execute())
+				})
+			})
+		})
+
+		when("env vars are passed as flags", func() {
+			var (
+				tmpVar   = "tmpVar"
+				tmpValue = "tmpKey"
+			)
 
 			it.Before(func() {
-				envfile, err := ioutil.TempFile("", "envfile")
-				h.AssertNil(t, err)
-				defer envfile.Close()
-
-				envfile.WriteString(`KEY=VALUE`)
-				envPath = envfile.Name()
+				h.AssertNil(t, os.Setenv(tmpVar, tmpValue))
 			})
 
 			it.After(func() {
-				h.AssertNil(t, os.RemoveAll(envPath))
+				h.AssertNil(t, os.Unsetenv(tmpVar))
 			})
 
-			it("builds an image env variables read from the env file", func() {
+			it("sets flag variables", func() {
 				mockClient.EXPECT().
 					Build(gomock.Any(), EqBuildOptionsWithEnv(map[string]string{
-						"KEY": "VALUE",
+						"KEY":  "VALUE",
+						tmpVar: tmpValue,
 					})).
 					Return(nil)
 
-				command.SetArgs([]string{"--builder", "my-builder", "image", "--env-file", envPath})
+				command.SetArgs([]string{"image", "--builder", "my-builder", "--env", "KEY=VALUE", "--env", tmpVar})
 				h.AssertNil(t, command.Execute())
 			})
 		})
 
-		when("two env files are provided with conflicted keys", func() {
-			var envPath1 string
-			var envPath2 string
-
-			it.Before(func() {
-				envfile1, err := ioutil.TempFile("", "envfile")
-				h.AssertNil(t, err)
-				defer envfile1.Close()
-
-				envfile1.WriteString("KEY1=VALUE1\nKEY2=IGNORED")
-				envPath1 = envfile1.Name()
-
-				envfile2, err := ioutil.TempFile("", "envfile")
-				h.AssertNil(t, err)
-				defer envfile2.Close()
-
-				envfile2.WriteString("KEY2=VALUE2")
-				envPath2 = envfile2.Name()
-			})
-
-			it.After(func() {
-				h.AssertNil(t, os.RemoveAll(envPath1))
-				h.AssertNil(t, os.RemoveAll(envPath2))
-			})
-
-			it("builds an image with the last value of each env variable", func() {
+		when("build fails", func() {
+			it("should show an error", func() {
 				mockClient.EXPECT().
-					Build(gomock.Any(), EqBuildOptionsWithEnv(map[string]string{
-						"KEY1": "VALUE1",
-						"KEY2": "VALUE2",
-					})).
-					Return(nil)
+					Build(gomock.Any(), gomock.Any()).
+					Return(errors.New(""))
 
-				command.SetArgs([]string{"--builder", "my-builder", "image", "--env-file", envPath1, "--env-file", envPath2})
-				h.AssertNil(t, command.Execute())
+				command.SetArgs([]string{"--builder", "my-builder", "image"})
+				err := command.Execute()
+				h.AssertError(t, err, "failed to build")
 			})
 		})
 
@@ -299,6 +420,148 @@ version = "1.0"
 					})
 				})
 			})
+
+			when("descriptor buildpack has uri", func() {
+				var projectTomlPath string
+
+				it.Before(func() {
+					projectToml, err := ioutil.TempFile("", "project.toml")
+					h.AssertNil(t, err)
+					defer projectToml.Close()
+
+					projectToml.WriteString(`
+[project]
+name = "Sample"
+
+[[build.buildpacks]]
+id = "example/lua"
+uri = "https://www.test.tgz"
+`)
+					projectTomlPath = projectToml.Name()
+				})
+
+				it.After(func() {
+					h.AssertNil(t, os.RemoveAll(projectTomlPath))
+				})
+
+				it("should build an image with configuration in descriptor", func() {
+					mockClient.EXPECT().
+						Build(gomock.Any(), EqBuildOptionsWithBuildpacks([]string{
+							"https://www.test.tgz",
+						})).
+						Return(nil)
+
+					command.SetArgs([]string{"image", "--builder", "my-builder", "--descriptor", projectTomlPath})
+					h.AssertNil(t, command.Execute())
+				})
+			})
+
+			when("descriptor buildpack has malformed uri", func() {
+				var projectTomlPath string
+
+				it.Before(func() {
+					projectToml, err := ioutil.TempFile("", "project.toml")
+					h.AssertNil(t, err)
+					defer projectToml.Close()
+
+					projectToml.WriteString(`
+[project]
+name = "Sample"
+
+[[build.buildpacks]]
+id = "example/lua"
+uri = "://bad-uri"
+`)
+					projectTomlPath = projectToml.Name()
+				})
+
+				it.After(func() {
+					h.AssertNil(t, os.RemoveAll(projectTomlPath))
+				})
+
+				it("should build an image with configuration in descriptor", func() {
+					mockClient.EXPECT().
+						Build(gomock.Any(), EqBuildOptionsWithBuildpacks([]string{
+							"https://www.test.tgz",
+						})).
+						Return(nil)
+
+					command.SetArgs([]string{"image", "--builder", "my-builder", "--descriptor", projectTomlPath})
+					err := command.Execute()
+					h.AssertError(t, err, "parse")
+				})
+			})
+
+			when("descriptor has exclude", func() {
+				var projectTomlPath string
+
+				it.Before(func() {
+					projectToml, err := ioutil.TempFile("", "project.toml")
+					h.AssertNil(t, err)
+					defer projectToml.Close()
+
+					projectToml.WriteString(`
+[project]
+name = "Sample"
+
+[build]
+exclude = [ "*.jar" ]
+`)
+					projectTomlPath = projectToml.Name()
+				})
+
+				it.After(func() {
+					h.AssertNil(t, os.RemoveAll(projectTomlPath))
+				})
+
+				it("should return appropriate fileFilter function", func() {
+					mockFilter := func(string) bool {
+						return false
+					}
+
+					mockClient.EXPECT().
+						Build(gomock.Any(), EqBuildOptionsWithFileFilter(mockFilter, "test.jar")).
+						Return(nil)
+
+					command.SetArgs([]string{"image", "--builder", "my-builder", "--descriptor", projectTomlPath})
+					h.AssertNil(t, command.Execute())
+				})
+			})
+
+			when("descriptor has include", func() {
+				var projectTomlPath string
+				it.Before(func() {
+					projectToml, err := ioutil.TempFile("", "project.toml")
+					h.AssertNil(t, err)
+					defer projectToml.Close()
+
+					projectToml.WriteString(`
+[project]
+name = "Sample"
+
+[build]
+include = [ "*.jar" ]
+`)
+					projectTomlPath = projectToml.Name()
+				})
+
+				it.After(func() {
+					h.AssertNil(t, os.RemoveAll(projectTomlPath))
+				})
+
+				it("should return appropriate fileFilter function", func() {
+					mockFilter := func(string) bool {
+						return true
+					}
+
+					mockClient.EXPECT().
+						Build(gomock.Any(), EqBuildOptionsWithFileFilter(mockFilter, "test.jar")).
+						Return(nil)
+
+					command.SetArgs([]string{"image", "--builder", "my-builder", "--descriptor", projectTomlPath})
+					h.AssertNil(t, command.Execute())
+				})
+			})
 		})
 	})
 }
@@ -326,6 +589,24 @@ func EqBuildOptionsWithNetwork(network string) gomock.Matcher {
 		description: fmt.Sprintf("Network=%s", network),
 		equals: func(o pack.BuildOptions) bool {
 			return o.ContainerConfig.Network == network
+		},
+	}
+}
+
+func EqBuildOptionsWithTrustedBuilder(trustBuilder bool) gomock.Matcher {
+	return buildOptionsMatcher{
+		description: fmt.Sprintf("Trust Builder=%t", trustBuilder),
+		equals: func(o pack.BuildOptions) bool {
+			return o.TrustBuilder == trustBuilder
+		},
+	}
+}
+
+func EqBuildOptionsWithFileFilter(fileFilter func(string) bool, fileName string) gomock.Matcher {
+	return buildOptionsMatcher{
+		description: fmt.Sprintf("File Filter=%p", fileFilter),
+		equals: func(o pack.BuildOptions) bool {
+			return o.FileFilter(fileName) == fileFilter(fileName)
 		},
 	}
 }
