@@ -7,10 +7,12 @@ import (
 	"os/signal"
 	"syscall"
 
+	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 
 	"github.com/buildpacks/pack"
 	"github.com/buildpacks/pack/internal/config"
+	"github.com/buildpacks/pack/internal/style"
 	"github.com/buildpacks/pack/logging"
 )
 
@@ -47,13 +49,25 @@ func logError(logger logging.Logger, f func(cmd *cobra.Command, args []string) e
 		cmd.SilenceUsage = true
 		err := f(cmd, args)
 		if err != nil {
-			if !IsSoftError(err) {
+			if _, isSoftError := errors.Cause(err).(pack.SoftError); !isSoftError {
 				logger.Error(err.Error())
+			}
+
+			if _, isExpError := errors.Cause(err).(pack.ExperimentError); isExpError {
+				configPath, err := config.DefaultConfigPath()
+				if err != nil {
+					return err
+				}
+				enableExperimentalTip(logger, configPath)
 			}
 			return err
 		}
 		return nil
 	}
+}
+
+func enableExperimentalTip(logger logging.Logger, configPath string) {
+	logging.Tip(logger, "To enable experimental features, add %s to %s.", style.Symbol("experimental = true"), style.Symbol(configPath))
 }
 
 func multiValueHelp(name string) string {
