@@ -10,7 +10,12 @@ import (
 	"github.com/buildpacks/pack/logging"
 )
 
+type CompletionFlags struct {
+	Shell string
+}
+
 func CompletionCommand(logger logging.Logger) *cobra.Command {
+	var flags CompletionFlags
 	var completionCmd = &cobra.Command{
 		Use:   "completion",
 		Short: "Outputs completion script location",
@@ -19,6 +24,11 @@ func CompletionCommand(logger logging.Logger) *cobra.Command {
 To configure your bash shell to load completions for each session, add the following to your '.bashrc' or '.bash_profile':
 
 	. $(pack completion)
+
+To configure your zsh shell to load completions for each session, add the following to your '.zshrc':
+
+	. $(pack completion --shell zsh)
+  
 	`,
 		RunE: logError(logger, func(cmd *cobra.Command, args []string) error {
 			packHome, err := config.PackHome()
@@ -31,13 +41,25 @@ To configure your bash shell to load completions for each session, add the follo
 			}
 			completionPath := filepath.Join(packHome, "completion")
 
-			if err := cmd.Parent().GenBashCompletionFile(completionPath); err != nil {
-				return err
+			var flagErr error
+			switch flags.Shell {
+			case "bash":
+				flagErr = cmd.Parent().GenBashCompletionFile(completionPath)
+			case "zsh":
+				flagErr = cmd.Parent().GenZshCompletionFile(completionPath)
+			default:
+				return errors.Errorf("%s is unsupported shell", flags.Shell)
+			}
+
+			if flagErr != nil {
+				return flagErr
 			}
 
 			logger.Info(completionPath)
 			return nil
 		}),
 	}
+
+	completionCmd.Flags().StringVarP(&flags.Shell, "shell", "s", "bash", "Generates completion file for [bash|zsh]")
 	return completionCmd
 }
