@@ -8,6 +8,7 @@ import (
 
 	"github.com/buildpacks/pack"
 	"github.com/buildpacks/pack/builder"
+	"github.com/buildpacks/pack/internal/config"
 	"github.com/buildpacks/pack/internal/style"
 	"github.com/buildpacks/pack/logging"
 )
@@ -16,15 +17,32 @@ type CreateBuilderFlags struct {
 	BuilderTomlPath string
 	Publish         bool
 	NoPull          bool
+	Registry        string
 }
 
-func CreateBuilder(logger logging.Logger, client PackClient) *cobra.Command {
+func validateCreateBuilderFlags(flags CreateBuilderFlags, cfg config.Config) error {
+	if flags.Publish && flags.NoPull {
+		return errors.Errorf("The --publish and --no-pull flags cannot be used together. The --publish flag requires the use of remote images.")
+	}
+
+	if flags.Registry != "" && !cfg.Experimental {
+		return pack.NewExperimentError("Support for buildpack registries is currently experimental.")
+	}
+
+	return nil
+}
+
+func CreateBuilder(logger logging.Logger, cfg config.Config, client PackClient) *cobra.Command {
 	var flags CreateBuilderFlags
 	cmd := &cobra.Command{
 		Use:   "create-builder <image-name> --builder-config <builder-config-path>",
 		Args:  cobra.ExactArgs(1),
 		Short: "Create builder image",
 		RunE: logError(logger, func(cmd *cobra.Command, args []string) error {
+			if err := validateCreateBuilderFlags(flags, cfg); err != nil {
+				return err
+			}
+
 			builderConfig, warns, err := builder.ReadConfig(flags.BuilderTomlPath)
 			if err != nil {
 				return errors.Wrap(err, "invalid builder toml")
@@ -39,6 +57,7 @@ func CreateBuilder(logger logging.Logger, client PackClient) *cobra.Command {
 				Config:      builderConfig,
 				Publish:     flags.Publish,
 				NoPull:      flags.NoPull,
+				Registry:    flags.Registry,
 			}); err != nil {
 				return err
 			}
@@ -49,6 +68,10 @@ func CreateBuilder(logger logging.Logger, client PackClient) *cobra.Command {
 	}
 	cmd.Flags().BoolVar(&flags.NoPull, "no-pull", false, "Skip pulling build image before use")
 	cmd.Flags().StringVarP(&flags.BuilderTomlPath, "builder-config", "b", "", "Path to builder TOML file (required)")
+	cmd.Flags().StringVarP(&flags.Registry, "buildpack-registry", "R", cfg.DefaultRegistry, "Buildpack Registry URL")
+	if !cfg.Experimental {
+		cmd.Flags().MarkHidden("buildpack-registry")
+	}
 	cmd.MarkFlagRequired("builder-config")
 	cmd.Flags().BoolVar(&flags.Publish, "publish", false, "Publish to registry")
 	AddHelpFlag(cmd, "create-builder")
