@@ -9,13 +9,20 @@ import (
 )
 
 type Config struct {
-	RunImages      []RunImage `toml:"run-images"`
-	DefaultBuilder string     `toml:"default-builder-image,omitempty"`
+	RunImages       []RunImage       `toml:"run-images"`
+	DefaultBuilder  string           `toml:"default-builder-image,omitempty"`
+	DefaultRegistry string           `toml:"default-registry-url,omitempty"`
+	Experimental    bool             `toml:"experimental,omitempty"`
+	TrustedBuilders []TrustedBuilder `toml:"trusted-builders,omitempty"`
 }
 
 type RunImage struct {
 	Image   string   `toml:"image"`
 	Mirrors []string `toml:"mirrors"`
+}
+
+type TrustedBuilder struct {
+	Name string `toml:"name"`
 }
 
 func DefaultConfigPath() (string, error) {
