@@ -7,7 +7,6 @@ import (
 	"testing"
 
 	"github.com/heroku/color"
-
 	"github.com/sclevine/spec"
 	"github.com/sclevine/spec/report"
 
@@ -46,6 +45,7 @@ func testConfig(t *testing.T, when spec.G, it spec.S) {
 				h.AssertNil(t, err)
 				h.AssertEq(t, subject.DefaultBuilder, "")
 				h.AssertEq(t, len(subject.RunImages), 0)
+				h.AssertEq(t, subject.Experimental, false)
 			})
 		})
 	})
@@ -65,6 +65,9 @@ func testConfig(t *testing.T, when spec.G, it spec.S) {
 							Mirrors: []string{"example.com/other/run", "example.com/other/mirror"},
 						},
 					},
+					TrustedBuilders: []config.TrustedBuilder{
+						{Name: "some-trusted-builder"},
+					},
 				}, configPath))
 				b, err := ioutil.ReadFile(configPath)
 				h.AssertNil(t, err)
@@ -76,6 +79,9 @@ func testConfig(t *testing.T, when spec.G, it spec.S) {
 				h.AssertContains(t, string(b), `[[run-images]]
   image = "other/run"
   mirrors = ["example.com/other/run", "example.com/other/mirror"]`)
+
+				h.AssertContains(t, string(b), `[[trusted-builders]]
+  name = "some-trusted-builder"`)
 			})
 		})
 
